fix(Readmore): added scrolling to div and improved readme (#50864, #50846)
