+++ src/components/Slider/Slider.stories.js
@@ -12,14 +12,9 @@ export default {
   },
 };
 
-export const RegularSlider = (args) => (
-  <div style={{ marginTop: '2rem' }}>
-    <Slider id="slider2" {...args} />
-  </div>
-);
+export const RegularSlider = (args) => <Slider id="slider2" />;
 
 RegularSlider.args = {
-  children: 'Slider',
   name: 'Form item name (name)',
   inputType: 'number',
   value: 50,
@@ -38,7 +33,6 @@ export const hideTextInput = (args) => (
 );
 
 hideTextInput.args = {
-  children: 'Slider',
   name: 'Form item name (name)',
   inputType: 'number',
   hideTextInput: true,
@@ -67,7 +61,6 @@ export const Disabled = (args) => (
 );
 
 Disabled.args = {
-  children: 'Slider',
   name: 'Form item name (name)',
   inputType: 'number',
   disabled: true,
@@ -95,7 +88,6 @@ export const Fullwidth = (args) => (
 );
 
 Fullwidth.args = {
-  children: 'Slider',
   name: 'Form item name (name)',
   inputType: 'number',
   fullWidth: true,
