+++ src/documentation/BrowserWarning/BrowserWarning.stories.js
@@ -16,6 +16,7 @@ export default {
     componentSubtitle: 'Template',
     status: 'legacy',
     mdx: markdown,
+    previewWidth: 'full',
   },
 };
