+++ src/documentation/Intro/_intro.scss
@@ -21,7 +21,6 @@ $easing: cubic-bezier(0.17, 0.89, 0.32, 1.1); //ease-in-out;
     margin-top: -7%;
     padding-top: 10%;
   }
-  z-index: -1;
   position: relative;
   /*color: var(--ui-01);
   .wfp--link {
