+++ src/documentation/LoginPage/LoginPage.stories.js
@@ -16,6 +16,7 @@ export default {
     mdx: markdown,
     introText:
       'Login page for legacy applications not using WFPs single sign-on',
+    previewWidth: 'full',
   },
 };
