+++ src/documentation/RegularPage/RegularPage.stories.js
@@ -16,6 +16,7 @@ export default {
     mdx: markdown,
     introText:
       'The regular page template is used for content pages showing mainly linear texts like blog articles, wikis or news. Avoid using it for dashboards with a lot of Modules.',
+    previewWidth: 'full',
   },
 };
